feat: rewrite quickfact documentation
